requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
bleach==4.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==3.1.6
django-markdownify==0.9.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
Markdown==3.3.6
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
requests==2.27.1
sqlparse==0.4.2
webencodings==0.5.1
zipp==3.7.0